+++ test/text/test_write.py
@@ -57,7 +57,6 @@ def test_write_font_stretching(tmp_path):  # issue #478
     pdf.add_page()
     # built-in font
     pdf.set_font("Helvetica", "", 8)
-    pdf.set_fill_color(255, 255, 0)
     pdf.set_right_margin(pdf.w - right_boundary)
     pdf.write(txt=LOREM_IPSUM[:100])
     pdf.ln()
@@ -70,7 +69,6 @@ def test_write_font_stretching(tmp_path):  # issue #478
     pdf.set_stretching(100)
     pdf.add_font(fname=FONTS_DIR / "DroidSansFallback.ttf")
     pdf.set_font("DroidSansFallback", "", 8)
-    pdf.set_fill_color(255, 255, 0)
     pdf.write(txt=LOREM_IPSUM[:100])
     pdf.ln()
     pdf.ln()
@@ -129,3 +127,26 @@ def test_write_superscript(tmp_path):
     pdf.denom_lift = 1.0
     write_this()
     assert_pdf_equal(pdf, HERE / "write_superscript.pdf", tmp_path)
+
+
+def test_write_char_wrap(tmp_path):  # issue #649
+    right_boundary = 50
+    pdf = fpdf.FPDF()
+    pdf.add_page()
+    pdf.set_right_margin(pdf.w - right_boundary)
+    pdf.set_font("Helvetica", "", 10)
+    pdf.write(txt=LOREM_IPSUM[:200])
+    pdf.ln()
+    pdf.ln()
+    pdf.write(txt=LOREM_IPSUM[:200], wrapmode="CHAR")
+    pdf.ln()
+    pdf.ln()
+    pdf.set_font("Courier", "", 10)
+    txt = "     " + "abcdefghijklmnopqrstuvwxyz" * 3
+    pdf.write(txt=txt)
+    pdf.ln()
+    pdf.ln()
+    pdf.write(txt=txt, wrapmode="CHAR")
+    pdf.line(pdf.l_margin, 10, pdf.l_margin, 130)
+    pdf.line(right_boundary, 10, right_boundary, 130)
+    assert_pdf_equal(pdf, HERE / "write_char_wrap.pdf", tmp_path)